hfile bootstrap index de hadoop
