+++ hudi-common/src/main/java/org/apache/hudi/io/storage/HoodieAvroFileReaderFactory.java
@@ -20,14 +20,13 @@ package org.apache.hudi.io.storage;
 
 import org.apache.hudi.common.config.HoodieConfig;
 import org.apache.hudi.common.util.Option;
+import org.apache.hudi.common.util.ReflectionUtils;
+import org.apache.hudi.exception.HoodieException;
 import org.apache.hudi.storage.HoodieStorage;
-import org.apache.hudi.storage.HoodieStorageUtils;
 import org.apache.hudi.storage.StorageConfiguration;
 import org.apache.hudi.storage.StoragePath;
 
 import org.apache.avro.Schema;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.hbase.io.hfile.CacheConfig;
 
 import java.io.IOException;
 
@@ -45,11 +44,16 @@ public class HoodieAvroFileReaderFactory extends HoodieFileReaderFactory {
     if (isUseNativeHFileReaderEnabled(hoodieConfig)) {
       return new HoodieNativeAvroHFileReader(conf, path, schemaOption);
     }
-    CacheConfig cacheConfig = new CacheConfig(conf.unwrapAs(Configuration.class));
-    if (schemaOption.isPresent()) {
-      return new HoodieHBaseAvroHFileReader(conf, path, cacheConfig, HoodieStorageUtils.getStorage(path, conf), schemaOption);
+    try {
+      if (schemaOption.isPresent()) {
+        return (HoodieFileReader) ReflectionUtils.loadClass("org.apache.hudi.io.storage.HoodieHBaseAvroHFileReader",
+            new Class<?>[] {StorageConfiguration.class, StoragePath.class, Option.class}, conf, path, schemaOption);
+      }
+      return (HoodieFileReader) ReflectionUtils.loadClass("org.apache.hudi.io.storage.HoodieHBaseAvroHFileReader",
+          new Class<?>[] {StorageConfiguration.class, StoragePath.class}, conf, path);
+    } catch (HoodieException e) {
+      throw (IOException) e.getCause().getCause();
     }
-    return new HoodieHBaseAvroHFileReader(conf, path, cacheConfig);
   }
 
   @Override
@@ -63,8 +67,13 @@ public class HoodieAvroFileReaderFactory extends HoodieFileReaderFactory {
     if (isUseNativeHFileReaderEnabled(hoodieConfig)) {
       return new HoodieNativeAvroHFileReader(conf, content, schemaOption);
     }
-    CacheConfig cacheConfig = new CacheConfig(conf.unwrapAs(Configuration.class));
-    return new HoodieHBaseAvroHFileReader(conf, path, cacheConfig, storage, content, schemaOption);
+    try {
+      return (HoodieFileReader) ReflectionUtils.loadClass("org.apache.hudi.io.storage.HoodieHBaseAvroHFileReader",
+          new Class<?>[] {StorageConfiguration.class, StoragePath.class, HoodieStorage.class, byte[].class, Option.class},
+          conf, path, storage, content, schemaOption);
+    } catch (HoodieException e) {
+      throw (IOException) e.getCause().getCause();
+    }
   }
 
   @Override
